% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exit.R
\name{exit}
\alias{exit}
\title{Exit from R}
\usage{
exit(...)
}
\arguments{
\item{...}{Arguments passed to \code{quit()}}
}
\description{
Exit from R by typing \code{exit} or \code{exit()}
}
\examples{
\donttest{exit("no", 0)}
}
